Create a new module: `auth::username_token`
As we'll have other auth types in `mod auth`.
